requirements.txt
pygame-ce>=2.5.0
moderngl>=5.12.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0 